requirements.txt
Django>=4.2,<5.0
psycopg[binary]>=3.1
python-dotenv>=1.0.0
django-tinymce